Service Client Functions

ROS.call

ROS.exists

ROS.getService

ROS.isPersistent

ROS.isValid

ROS.call — Method

call(serviceClient, srv)

Call the service aliased by this handle with the specified service request/response.

ROS.exists — Method

exists(serviceClient)

Checks if this is both advertised and available.

ROS.getService — Method

getService(serviceClient)

Returns the name of the service this ServiceClient connects to.

ROS.isPersistent — Method

isPersistent(serviceClient)

Returns true if this handle points to a persistent service, false otherwise.

ROS.isValid — Method

isValid(serviceClient)

Returns whether or not this handle is valid. For a persistent service, this becomes false when the connection has dropped. Non-persistent service handles are always valid.