Time Functions

ROS.Duration

ROS.Duration

ROS.Duration

ROS.Duration

ROS.Rate

ROS.Rate

ROS.Time

ROS.Time

ROS.Time

ROS.cycleTime

ROS.expectedCycleTime

ROS.init

ROS.isSimTime

ROS.isSystemTime

ROS.isValid

ROS.now

ROS.reset

ROS.setNow

ROS.shutdown

ROS.sleep

ROS.sleep

ROS.sleepUntil

ROS.toSec

ROS.toSec

ROS.useSystemTime

ROS.waitForValid

ROS.Duration — Method

Duration(sec::Int, nsec::Int)

Constructor, creates a Duration.

ROS.Duration — Method

Duration(t::Number)

Constructor, creates a Duration.

ROS.Duration — Method

Duration(rate)

Constructor, creates a Duration from a Rate.

ROS.Duration — Method

Duration()

Default constructor, creates a Duration.

ROS.Rate — Method

Rate(frequency::Number)

Constructor, creates a Rate.

ROS.Rate — Method

Rate(duration)

Constructor, creates a Rate from a Duration.

ROS.Time — Method

Time(sec::Int, ncsec::Int)

Constructor, creates Time.

ROS.Time — Method

Time(t::Number)

Constructor, creates Time.

ROS.Time — Method

Time()

Default constructor, creates Time.

ROS.cycleTime — Method

cycleTime(rate)

Get the actual run time of a cycle from start to sleep.

ROS.expectedCycleTime — Method

expectedCycleTime(rate)

Get the expected cycle time – one over the frequency passed in to the constructor.

ROS.init — Method

init(time)

Initialise Time.

ROS.isSimTime — Method

isSimTime(time)

Returns whether or not sim time is used.

ROS.isSystemTime — Method

isSystemTime(time)

Returns whether or not system time is used.

ROS.isValid — Method

isValid(time)

Returns whether or not the current time source is valid. Simulation time is valid if it is non-zero.

ROS.now — Method

now(time)

Retrieve the current time. If ROS clock time is in use, this returns the time according to the ROS clock. Otherwise returns the current wall clock time.

ROS.reset — Method

reset(rate)

Sets the start time for the rate to now.

ROS.setNow — Method

setNow(time, now) = @cxx t->setNow(now)

Set now to now time.

ROS.shutdown — Method

shutdown(time)

Shuts down the Time.

ROS.sleep — Method

sleep(duration)

Sleep for the amount of time specified by this Duration. If a signal interrupts the sleep, resleeps for the time remaining.

ROS.sleep — Method

sleep(rate)

Sleeps for any leftover time in a cycle. Calculated from the last time sleep, reset, or the constructor was called.

ROS.sleepUntil — Method

sleepUntil(time, _end)

Sleep until a specific time has been reached.

ROS.toSec — Method

toSec(duration)

Turn Duration into floating point seconds.

ROS.toSec — Method

toSec(time)

Turn Time into floating point seconds.

ROS.useSystemTime — Method

useSystemTime(time)

Switch to system time.

ROS.waitForValid — Method

waitForValid(time)

Wait for time source to become valid.