TF Functions

ROS.tf.Buffer

ROS.tf.StaticTransformBroadcaster

ROS.tf.TransformBroadcaster

ROS.tf.TransformListener

ROS.tf.angle

ROS.tf.angleShortestPath

ROS.tf.canTransform

ROS.tf.canTransform

ROS.tf.dot

ROS.tf.getAngle

ROS.tf.getAngleShortestPath

ROS.tf.getAxis

ROS.tf.inverse

ROS.tf.length

ROS.tf.length2

ROS.tf.lookupTransform

ROS.tf.lookupTransform

ROS.tf.normalize

ROS.tf.sendTransform

ROS.tf.sendTransform

ROS.tf.sendTransform

ROS.tf.sendTransform

ROS.tf.setEuler

ROS.tf.setRPY

ROS.tf.setRotation

ROS.tf.slerp

ROS.tf.transform

ROS.tf.transform

ROS.tf.Buffer — Method

Buffer()

Constructor for a Buffer object.

ROS.tf.StaticTransformBroadcaster — Method

StaticTransformBroadcaster()

Constructor (needs a ros::Node reference).

ROS.tf.TransformBroadcaster — Method

TransformBroadcaster()

Constructor (needs a ros::Node reference).

ROS.tf.TransformListener — Method

TransformListener(buffer)

Constructor for transform listener.

ROS.tf.angle — Method

angle(q1, q2)

Return the half angle between q1 and q2 geometry_msgs::Quaternions.

ROS.tf.angleShortestPath — Method

angleShortestPath(q1, q2)

Return the angle between q1 and q2 geometry_msgs::Quaternions.

ROS.tf.canTransform — Method

canTransform(buffer, target_frame::String, target_time, source_frame::String, source_time, fixed_frame::String, timeout)

Test if a transform is possible.

ROS.tf.canTransform — Method

canTransform(buffer, target_frame::String, source_frame::String, time, timeout)

Test if a transform is possible.

ROS.tf.dot — Method

dot(q1, q2)

Return the dot product between q1 and q2 geometry_msgs::Quaternions.

ROS.tf.getAngle — Method

getAngle(quaternion)

Return the angle [0, 2π] of rotation represented by this quaternion.

ROS.tf.getAngleShortestPath — Method

getAngleShortestPath(quaternion)

Return the angle [0, π] of rotation represented by this quaternion.

ROS.tf.getAxis — Method

getAxis(quaternion)

Return the axis of the rotation represented by this geometry_msgs::Quaternion.

ROS.tf.inverse — Method

inverse(quaternion)

Return the inverse of this geometry_msgs::Quaternion.

ROS.tf.length — Method

length(quaternion)

Return the length of the quaternion.

ROS.tf.length2 — Method

length2(quaternion)

Return the length squared of the quaternion.

ROS.tf.lookupTransform — Method

lookupTransform(buffer, target_frame::String, target_time, source_frame::String, source_time, fixed_frame::String, timeout)

Get the transform between two frames by frame ID assuming fixed frame.

ROS.tf.lookupTransform — Method

lookupTransform(buffer, target_frame::String, source_frame::String, time, timeout)

Get the transform between two frames by frame ID.

ROS.tf.normalize — Method

normalize(quaternion)

Normalize a geometry_msgs::Quaternion (useful after calculations).

ROS.tf.sendTransform — Method

sendTransform(static_broadcaster, transforms::Union{Array,Vector})

Send a TransformStamped message The stamped data structure includes frameid, and time, and parentid already.

ROS.tf.sendTransform — Method

sendTransform(static_broadcaster, transform)

Send a TransformStamped message The stamped data structure includes frameid, and time, and parentid already.

ROS.tf.sendTransform — Method

function sendTransform(broadcaster, transforms::Union{Array,Vector})

Send a TransformStamped message The stamped data structure includes frameid, and time, and parentid already.

ROS.tf.sendTransform — Method

sendTransform(broadcaster, transform)

Send a TransformStamped message The stamped data structure includes frameid, and time, and parentid already.

ROS.tf.setEuler — Method

setEuler(yaw::Number, pitch::Number, roll::Number)

Euler angles (yaw/pitch/roll) to geometry_msgs::Quaternion.

ROS.tf.setRPY — Method

setRPY(roll::Number, pitch::Number, yaw::Number)

Roll,Pitch,Yaw to geometry_msgs::Quaternion.

ROS.tf.setRotation — Method

setRotation(quaternion, axis::Union{Array,Vector}, angle::Number)

Set the rotation using axis angle notation.

ROS.tf.slerp — Method

slerp(q1, q2, t::Number)

Return the geometrymsgs::Quaternion which is the result of Spherical Linear Interpolation between q1 and q2 geometrymsgs::Quaternions.

ROS.tf.transform — Method

transform(buffer, in, target_frame::String, timeout)

Transform an input into the target frame.

ROS.tf.transform — Method

transform(buffer, in, target_frame::String, target_time, fixed_frame::String, timeout)

Transform an input into the target frame.